api/requirements.txt
azure-functions>=1.21.0
azure-cosmos>=4.7.0
pydantic>=2.0.0
